[impact driver sound, hot glue smell] DO NOT USE
